Add the ability to dump some messages for debugging to the duplicator
This allows us to do `kill -USR1 <pid>` and force the duplicator to dump
100 messages to stdout.  This is useful for debugging live systems,
especially in the cloud where inserting debugger breakpoints isnʼt
feasible.

Probably needs to be performance benchmarked.
