+++ mite/__main__.py
@@ -71,6 +71,7 @@ import uvloop
 import msgpack
 
 from .cli.common import _create_config_manager
+from .cli.duplicator import duplicator
 from .cli.stats import stats
 from .collector import Collector
 from .controller import Controller
@@ -119,12 +120,6 @@ def _create_runner_transport(opts):
 def _create_controller_server(opts):
     socket = opts['--controller-socket']
     return _msg_backend_module(opts).ControllerServer(socket)
-
-
-def _create_duplicator(opts):
-    return _msg_backend_module(opts).Duplicator(
-        opts['--message-socket'], opts['OUT_SOCKET']
-    )
 
 
 logger = logging.getLogger(__name__)
@@ -398,11 +393,6 @@ def recorder(opts):
     recorder = Recorder(opts['--recorder-dir'])
     receiver.add_listener(recorder.process_message)
     asyncio.get_event_loop().run_until_complete(receiver.run())
-
-
-def duplicator(opts):
-    duplicator = _create_duplicator(opts)
-    asyncio.get_event_loop().run_until_complete(duplicator.run())
 
 
 def cat(opts):
